+++ internal/commands/all/resource.go
@@ -17,6 +17,8 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/networkpeering"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/objectstorage"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/router"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/server"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/storage"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/resolver"
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 )
@@ -30,6 +32,8 @@ const (
 	typeRouter         = "router"
 	typeObjectStorage  = "object-storage"
 	typeDatabase       = "database"
+	typeServer         = "server"
+	typeStorage        = "storage"
 )
 
 type Resource struct {
@@ -128,7 +132,7 @@ func findResources[T any](exec commands.Executor, wg *sync.WaitGroup, returnChan
 
 func listResources(exec commands.Executor, include, exclude []string) ([]Resource, error) {
 	var resources []Resource
-	returnChan := make(chan findResult, 5)
+	returnChan := make(chan findResult, 7)
 
 	var wg sync.WaitGroup
 
@@ -137,6 +141,8 @@ func listResources(exec commands.Executor, include, exclude []string) ([]Resourc
 	findResources(exec, &wg, returnChan, &resolver.CachingRouter{}, include, exclude)
 	findResources(exec, &wg, returnChan, &resolver.CachingObjectStorage{}, include, exclude)
 	findResources(exec, &wg, returnChan, &resolver.CachingDatabase{}, include, exclude)
+	findResources(exec, &wg, returnChan, &resolver.CachingServer{}, include, exclude)
+	findResources(exec, &wg, returnChan, &resolver.CachingStorage{}, include, exclude)
 
 	wg.Wait()
 	close(returnChan)
@@ -191,6 +197,18 @@ func getResource(val any) (Resource, error) {
 			Type: typeDatabase,
 			UUID: v.UUID,
 		}, nil
+	case upcloud.Server:
+		return Resource{
+			Name: v.Title,
+			Type: typeServer,
+			UUID: v.UUID,
+		}, nil
+	case upcloud.Storage:
+		return Resource{
+			Name: v.Title,
+			Type: typeStorage,
+			UUID: v.UUID,
+		}, nil
 	}
 	return Resource{}, fmt.Errorf("unsupported type %T", val)
 }
@@ -207,6 +225,10 @@ func deleteResource(exec commands.Executor, resource Resource) (err error) {
 		_, err = objectstorage.Delete(exec, resource.UUID, true, true, true, true)
 	case typeDatabase:
 		_, err = database.Delete(exec, resource.UUID, true, true)
+	case typeServer:
+		_, err = server.Delete(exec, resource.UUID, false, true)
+	case typeStorage:
+		_, err = storage.Delete(exec, resource.UUID, "delete")
 	}
 	return
 }