+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support server and storage resources in `all list` and `all purge` commands.
+- Add `--stop` flag to `server delete` command.
+- Add `--backups` flag to `storage delete` command.
+
 ## [3.18.0] - 2025-04-04
 
 ### Added